% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_classi.R
\name{knn_classi}
\alias{knn_classi}
\title{knn Class to predict}
\usage{
knn_classi(train_data, obse, type, calcols, classcol, idcol, k)
}
\arguments{
\item{train_data}{}

\item{obse}{}

\item{type}{}

\item{calcols}{}
}
\value{
a random sample of size
}
\description{
Use Knn to predict the category a point belongs to
}
