% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Two_Fold_Cross_Validation.R
\name{Two_Fold_Cross_Validation}
\alias{Two_Fold_Cross_Validation}
\title{Two_Fold Cross_Validation of Linear....}
\usage{
Two_Fold_Cross_Validation(a, b)
}
\arguments{
\item{a}{Forecast variables}

\item{b}{Response variables}
}
\value{
residual matrices
}
\description{
Two_Fold Cross_Validation of Linear
}
\examples{
\dontrun{
data(data)
attach(data)
re <- Two_Fold_Cross_Validation(a,b)
print(re)
}
}
